+++ example/lib/pages/mini_games/simple_example/simple_example_game.dart
@@ -1,8 +1,8 @@
 import 'package:bonfire/bonfire.dart';
-import 'package:example/shared/decoration/barrel_dragable.dart';
-import 'package:example/shared/decoration/potion_life.dart';
 import 'package:example/pages/mini_games/simple_example/my_enemy.dart';
 import 'package:example/pages/mini_games/simple_example/my_player.dart';
+import 'package:example/shared/decoration/barrel_dragable.dart';
+import 'package:example/shared/decoration/potion_life.dart';
 import 'package:flutter/material.dart';
 
 ///