+++ lib/util/direction.dart
@@ -11,23 +11,29 @@ enum Direction {
   bool get isVertical => this == down || this == up;
   bool get isHorizontal => this == left || this == right;
 
+  bool get isLeftSide => this == left || this == upLeft || this == downLeft;
+  bool get isRightSide => this == right || this == upRight || this == downRight;
+
+  bool get isUpSide => this == up || this == upRight || this == upLeft;
+  bool get isDownSide => this == down || this == downRight || this == downLeft;
+
   bool isSameXDirection(double x) {
-    if (x > 0 && this == right) {
+    if (x > 0 && (this == right || this == upRight || this == downRight)) {
       return true;
     }
 
-    if (x < 0 && this == left) {
+    if (x < 0 && (this == left || this == upLeft || this == downLeft)) {
       return true;
     }
     return false;
   }
 
   bool isSameYDirection(double y) {
-    if (y > 0 && this == down) {
+    if (y > 0 && (this == down || this == downRight || this == downLeft)) {
       return true;
     }
 
-    if (y < 0 && this == up) {
+    if (y < 0 && (this == up || this == upRight || this == upLeft)) {
       return true;
     }
     return false;