+++ lib/util/value_generator_component.dart
@@ -10,22 +10,27 @@ class ValueGeneratorComponent extends Component {
   final double begin;
   final double end;
   final Curve curve;
+  final Curve? reverseCurve;
   final VoidCallback? onFinish;
   final ValueChanged<double>? onChange;
   final bool autoStart;
+  final bool infinite;
 
   double _currentValue = 0;
   double _displacement = 0;
   bool _isRunning = false;
+  bool _reversing = false;
 
   ValueGeneratorComponent(
     this.duration, {
     this.begin = 0,
     this.end = 1,
     this.curve = Curves.decelerate,
+    this.reverseCurve,
     this.onFinish,
     this.onChange,
-    this.autoStart = false,
+    this.autoStart = true,
+    this.infinite = false,
   }) {
     _isRunning = autoStart;
     _displacement = end - begin;
@@ -35,12 +40,32 @@ class ValueGeneratorComponent extends Component {
   void updateTree(double dt) {
     if (!_isRunning) return;
 
-    _currentValue += dt * _maxInMilliSeconds;
-    if (_currentValue >= duration.inMilliseconds) {
-      finish();
+    if (_reversing) {
+      _currentValue -= dt * _maxInMilliSeconds;
     } else {
-      double value = curve.transform(_currentValue / duration.inMilliseconds);
-      double realValue = begin + (_displacement * value);
+      _currentValue += dt * _maxInMilliSeconds;
+    }
+
+    if (_currentValue >= duration.inMilliseconds) {
+      if (infinite) {
+        _reversing = true;
+      } else {
+        finish();
+      }
+    } else if (_currentValue <= 0 && _reversing) {
+      _currentValue = 0;
+      _reversing = false;
+    } else {
+      double realValue;
+      if (_reversing) {
+        double value = (reverseCurve ?? curve)
+            .transform(1 - (_currentValue / duration.inMilliseconds));
+        realValue = begin + (_displacement * (1 - value));
+      } else {
+        double value = curve.transform(_currentValue / duration.inMilliseconds);
+        realValue = begin + (_displacement * value);
+      }
+
       onChange?.call(realValue);
     }
   }